Fix public key fetcher bug when the split public key has an "n" in the edges.
Public keys may be split in more than one TXT record. For such cases, there's a chance that the letter "n" sits exactly at the beginning or at the end of some of the record strings, we cannot use python's split('\\n') because it will take out the "n" and generate a wrong public key. Changed by replace('\\n', '') to avoid this issue.